+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ===================
 - Fix bug in append string when no data was present
 - Update default color scheme
+- Add search functionality. Trigger search via '/'.
 
 Released
 --------